:muscle: check before deleting

+++ scripts/install-nvim.sh
@@ -6,17 +6,17 @@ make CMAKE_BUILD_TYPE=Release CMAKE_INSTALL_PREFIX=~/.local/nvim install
 
 # Remove default plugins
 if [ -d ~/.local/nvim/share/nvim/runtime/plugin ]; then
-    rm ~/.local/nvim/share/nvim/runtime/plugin/gzip.vim
-    rm ~/.local/nvim/share/nvim/runtime/plugin/health.vim
-    rm ~/.local/nvim/share/nvim/runtime/plugin/matchit.vim
-    rm ~/.local/nvim/share/nvim/runtime/plugin/matchparen.vim
-    rm ~/.local/nvim/share/nvim/runtime/plugin/netrwPlugin.vim
-    rm ~/.local/nvim/share/nvim/runtime/plugin/shada.vim
-    rm ~/.local/nvim/share/nvim/runtime/plugin/spellfile.vim
-    rm ~/.local/nvim/share/nvim/runtime/plugin/tarPlugin.vim
-    rm ~/.local/nvim/share/nvim/runtime/plugin/tohtml.vim
-    rm ~/.local/nvim/share/nvim/runtime/plugin/tutor.vim
-    rm ~/.local/nvim/share/nvim/runtime/plugin/zipPlugin.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/gzip.vim ] && rm ~/.local/nvim/share/nvim/runtime/plugin/gzip.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/health.vim ] && rm ~/.local/nvim/share/nvim/runtime/plugin/health.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/matchit.vim ] && rm ~/.local/nvim/share/nvim/runtime/plugin/matchit.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/matchparen.vim ] && rm ~/.local/nvim/share/nvim/runtime/plugin/matchparen.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/netrwPlugin.vim ] && rm ~/.local/nvim/share/nvim/runtime/plugin/netrwPlugin.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/shada.vim ] && rm ~/.local/nvim/share/nvim/runtime/plugin/shada.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/spellfile.vim ] && rm ~/.local/nvim/share/nvim/runtime/plugin/spellfile.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/tarPlugin.vim ] && rm ~/.local/nvim/share/nvim/runtime/plugin/tarPlugin.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/tohtml.vim ] && rm ~/.local/nvim/share/nvim/runtime/plugin/tohtml.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/tutor.vim ] && rm ~/.local/nvim/share/nvim/runtime/plugin/tutor.vim
+    [ -f ~/.local/nvim/share/nvim/runtime/plugin/zipPlugin.vim] && rm ~/.local/nvim/share/nvim/runtime/plugin/zipPlugin.vim
 fi
 if [ -f ~/.local/nvim/share/nvim/runtime/ftplugin.vim ]; then
     rm ~/.local/nvim/share/nvim/runtime/ftplugin.vim
